pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mega-engine"
version = "1.1.0"
description = "Pipeline estatistico para geracao e avaliacao de jogos da Mega-Sena"
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "openai",
  "pandas",
  "pillow",
  "requests",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
